requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# CLI utilities
